AstroIMFs

Documentation for AstroIMFs.

AstroIMFs.AbstractIMF

AstroIMFs.PowerlawIMF

AstroIMFs.eval_by_mass

AstroIMFs.eval_by_num

AstroIMFs.integrate_eval_by_mass

AstroIMFs.integrate_eval_by_mass

AstroIMFs.integrate_eval_by_num

AstroIMFs.mass_in_massinterval

AstroIMFs.nstars_in_massinterval

AstroIMFs.AbstractIMF — Type

abstract type AbstractIMF end

The abstract type for stellar initial mass functions.

The following functions can be used on objects of this supertype:

eval_by_mass

eval_by_num

integrate_eval_by_mass

integrate_eval_by_num

mass_in_massinterval

nstars_in_massinterval

For developers

The following methods have to be defined for subtypes:

eval_by_mass

eval_by_num

AstroIMFs.PowerlawIMF — Type

struct PowerlawIMF{M,S,T} <: AbstractIMF
    massmin::M
    massmax::M
    slopemasses::Vector{M}
    slopes::Vector{S}
    slopenorms::Vector{T}
end

A powerlaw IMF with multiple separate powerlaw sections ranging the mass range from massmin to massmax.

This IMF can also handle unitful quantities, i.e. using solar masses as units.

Attributes

slopemasses: vector of edge masses between the slopes of length N, where the last element is equal to massmin (and the first element must be smaller or equal to massmax)

slopes: vector of the powerlaw segment slopes

slopenorms: vector of normalization values for the powerlaw segments to force the integral to be 1

Examples

# for a Chabrier powerlaw IMF:
imf_chabrier = PowerlawIMF(0.1, 100.0, [1, .3, .1], [1.3, 0.8, 0.2], [0.241367, 0.241367, 0.497056])

AstroIMFs.eval_by_mass — Method

eval_by_mass(imf::PowerlawIMF, m::Number)

Returns the IMF probability density function evaluated at the mass m.

Integrating this over all masses results in 1.

AstroIMFs.eval_by_num — Method

eval_by_num(imf::PowerlawIMF, m::Number)

Returns the IMF probability density function divided by the mass evaluated at the mass m.

Multiplying the integral of this over a mass interval with the total mass of stars that the IMF applies to results in the expected number of stars within that interval.

AstroIMFs.integrate_eval_by_mass — Method

integrate_eval_by_mass(imf::PowerlawIMF, mrange)

Return the integral over eval_by_mass for the mass range mrange as a Tuple or AbstractArray.

AstroIMFs.integrate_eval_by_mass — Method

integrate_eval_by_num(imf::PowerlawIMF, mrange)

Return the integral over eval_by_mass for the mass range mrange as a Tuple or AbstractArray.

AstroIMFs.integrate_eval_by_num — Method

integrate_eval_by_num(imf::PowerlawIMF, mrange)

Return the integral over eval_by_num for the mass range mrange as a Tuple or AbstractArray.

AstroIMFs.mass_in_massinterval — Method

mass_in_massinterval(imf::AbstractIMF, mtot::Number, mrange)

Return the expected total mass found within the mass range mrange for the imf of mass mtot.

The mass range is given as a Tuple or AbstractArray.

AstroIMFs.nstars_in_massinterval — Method

nstars_in_massinterval(imf::AbstractIMF, mtot::Number, mrange)

Return the expected total number of stars found within the mass range mrange for the imf of mass mtot.

The mass range is given as a Tuple or AbstractArray.